Fix `EnsureSingleSpaceAfterTokenChecker` when token is first one on line.
In this case the token's "associated whitespace" contains a newline. If checking for space _before_ token, then this means the check can be skipped - no need to check for a space before if there is a newline before. The bug is when the check is for space _after_: in this case we do _not_ want to skip the check - having a newline before does not mean we don't want a space after!