Fixes #48 PublicMethodsHaveTypeChecker should not flag nested defs
This fixes the erroneous message when a def is defined within a def.

def foo(): Int = { def foobar() = 66; foobar() }

The above should not cause an error. Scalastyle now considers any method defined within a def to be ignored, at any level. So the following will also not cause an error:

def foo(): Int = { class Foobar { def foobar() = 66 }; new Foobar().foobar() }

This may be considered to be valid, but if foobar() and Foobar are not available outside the method.