Less verbose creation of ScalastyleConfiguration
Decreasing the verboseness of `ScalastyleConfiguration` creation can be useful,
for example, to the substitution of the process of parsing the XML
configuration file by the manual creation of the object itself in Scala (which
is "type-safer") in the SBT plugin.
This adds default parameters to `ScalastyleConfiguration` and a new `apply` to
`ConfigurationChecker` that uses a `ClassTag` to get the class name.